+++ package.json
@@ -13,6 +13,7 @@
     "react-dom": "^16.13.1",
     "react-popper": "^2.2.3",
     "react-redux": "^7.2.1",
+    "react-redux-form": "^1.16.14",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "reactstrap": "^8.5.1",
